Found one of the aforementioned bugs. Putting the "one of the Parameters being merged must be the last element of a Route" was a stupid requirement, as indicated by this case:
	/:forum/post
	/:forum/:thread

Obviously, this is decidable despite neither of the Route fragments after the first parameter are RouteEnd's. So I removed that constraint (keeping the "parameter names must be the same" one) and just merge the two cdr's of a Parameter (which other code turns into a Branch fragment, etc).

Seems to work.