Remember that bit about "oh hey let's not write a whole bloody fucking ORM for this shit?" Well. I lied.
Going to define my schema in a quaint little Haskell table, then have an extra step in the build process which takes the table and spits out a schema definition, in addition to records and shit for each of the tables. It won't be an ORM like Ruby provides, but rather just an automated means to parse the SqlType's from HDBC fetchRow calls into records so I don't have to fiddle with that shit.

Don't necessarily have to use records, but meh they'll work. I want *something* that I can pass around that isn't opaque (fuck you, SqlTypes are opaque).