Added a couple of lines of code to the ORM layer to make DbRecord derive from Text.JSON.JSON (implemented only the encoder side -- don't give a shit about the decoding stuff). So now all the records in the ORM wrapper code generated from the schema can automatically be serialized to JSON which makes life a little bit easier.
Realistically, I'm beginning to think (once again) that this entire ORM layer is a terrible thing to have written. The fact that I'm using records (or, realistically, specialized data types of any kind) necessitates this kind of bloatcode; though going with something like a Data.Map.Map (String, SqlValue) necessitates unboxing the values with fromSql (or whatever).

Dunno which is better, but unboxing is a lot less work. Mostly.