Bloated the ORM layer up more (I really need to switch it over to use HStringTemplate to generate the code, lawds) by adding the following methods into DbRecord:
	toAscList :: DbRecord a => a -> [(String, String)]
	toAscListJSON :: DbRecord a => a -> [(String, JSValue)]

Which take records and convert them into associative lists with show and showJSON, respectively. The showJSON method for DbRecords has been simplified to just be `makeObj . toAscListJSON` -- the complexity of the function is still in the code, it's just been shuffled around a bit.

Now the records should play nicely with HStringTemplate. I have to poop.