Changed all the of the application functions to have the signature
	(IConnection a, Stringable b) => a -> STGroup b -> RouteParameters -> IO String

and removed all the type declarations from them because it was getting incredibly messy (my god). Changed LulzBB/Main.hs so that all of the values in the rawActionTable get curried properly with this added parameter (and added in the bit of code which loads all the templates from a directory). Should be all set now to start generating some HTML hur hur.